backend/services/ai_service.py
import cv2
import random

# TODO: Load actual YOLOv8 `.pt` model here
# from ultralytics import YOLO
# yolo_model = YOLO('models/yolov8_reference.pt')

# TODO: Load actual PyTorch CNN MobileNet/EfficientNet model here
# import torch
# cnn_model = torch.load('models/indian_food_cnn.pt')

KNOWN_COIN_DIAMETER_CM = 2.3  # E.g., Rs 5 coin
KNOWN_SPOON_LENGTH_CM = 15.0

def process_image(image_path):
    """
    Takes an image, runs YOLO to find reference object + food.
    Classifies the food using CNN.
    Calculates scale and food area in cm^2.
    """
    
    # Read image using OpenCV to verify it's valid
    img = cv2.imread(image_path)
    if img is None:
        raise ValueError("Invalid image file.")
        
    height, width, _ = img.shape
    
    # MOCK YOLOv8 DETECTION
    # In a real scenario: results = yolo_model(img)
    # Extract bounding boxes for coin/spoon and food
    
    # Mock finding a coin
    coin_pixel_diameter = width * 0.1  # Suppose coin is 10% of image width
    
    # Calculate scale: pixels to cm
    pixels_per_cm = coin_pixel_diameter / KNOWN_COIN_DIAMETER_CM
    
    # Mock finding food bounding box
    food_box_width_px = width * 0.4
    food_box_height_px = height * 0.3
    
    # Calculate real world area of food in cm^2
    food_width_cm = food_box_width_px / pixels_per_cm
    food_height_cm = food_box_height_px / pixels_per_cm
    area_cm2 = food_width_cm * food_height_cm
    
    # MOCK CNN CLASSIFICATION
    # In a real scenario: food_class = cnn_model.predict(cropped_food_img)
    possible_foods = ["idli", "dosa", "rice", "chapati"]
    food_name = random.choice(possible_foods)
    
    return {
        "food_name": food_name,
        "area_cm2": area_cm2,
        "scale_px_per_cm": pixels_per_cm
    }
