backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
pyjwt
ultralytics
torch
torchvision
opencv-python
numpy
python-dotenv
requests
